requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
email-validator==2.2.0
PyJWT==2.10.1
pytest==8.3.5
httpx==0.28.1
